+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/v2/internal/GenerateCustomerRecommendationsApi.kt
@@ -77,6 +77,7 @@ internal class GenerateCustomerRecommendationsApi(
                 val jsonRequestObjects = customerSessionRequestBuilder.createRequestObjects(customerSessionRequest)
                 put(CUSTOMER, jsonRequestObjects.customer)
                 putOpt(PURCHASE_UNITS, jsonRequestObjects.purchaseUnits)
+                putOpt(PAYPAL_CAMPAIGNS, jsonRequestObjects.campaigns)
             }
         }
 
@@ -125,6 +126,7 @@ internal class GenerateCustomerRecommendationsApi(
         private const val SESSION_ID = "sessionId"
         private const val CUSTOMER = "customer"
         private const val PURCHASE_UNITS = "purchaseUnits"
+        private const val PAYPAL_CAMPAIGNS = "paypalCampaigns"
         private const val GENERATE_CUSTOMER_RECOMMENDATIONS = "generateCustomerRecommendations"
     }
 }